# openapi/openapi-baseline.json
{
  "components": {
    "schemas": {
      "ActionPlanActionRequest": {
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/ActionPlanActivityRequest"
            },
            "minItems": 1,
            "type": "array"
          },
          "needId": {
            "format": "uuid",
            "type": "string"
          },
          "outcomeId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "activities",
          "needId",
          "outcomeId"
        ],
        "type": "object"
      },
      "ActionPlanActionResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "message",
          "success"
        ],
        "type": "object"
      },
      "ActionPlanActivityRequest": {
        "properties": {
          "activityDetails": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "minLength": 1,
            "type": "string"
          },
          "who": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "activityDetails",
          "status",
          "who"
        ],
        "type": "object"
      },
...
      "ActionPlanSummaryDto": {
        "properties": {
          "needs": {
            "items": {
              "$ref": "#/components/schemas/ActionPlanSummaryNeed"
            },
            "type": "array"
          },
          "personDetails": {
            "$ref": "#/components/schemas/ActionPlanSummaryPersonDetails"
          }
        },
        "required": [
          "needs",
          "personDetails"
        ],
        "type": "object"
      },
      "ActionPlanSummaryNeed": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/ActionPlanSummaryOutcome"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "label",
          "outcomes"
        ],
        "type": "object"
      },
      "ActionPlanSummaryOutcome": {
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/ActionPlanSummaryOutcomeActivity"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "activities",
          "id",
          "label"
        ],
        "type": "object"
      },
      "ActionPlanSummaryOutcomeActivity": {
        "properties": {
          "details": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "who": {
            "type": "string"
          }
        },
        "required": [
          "details",
          "id",
          "who"
        ],
        "type": "object"
      },
      "ActionPlanSummaryPersonDetails": {
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "email": "feedback@digital.justice.gov.uk",
      "name": "HMPPS Digital Studio"
    },
    "title": "HMPPS Community Support Api"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bff/referral/{referralReference}/action-plan": {
      "get": {
        "operationId": "getActionPlanSummary",
        "parameters": [
          {
            "in": "path",
            "name": "referralReference",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlanSummaryDto"
                }
              }
            },
            "description": "Action Plan, and Referral, found - data returned"
          },
          "404": {
            "content": {
              "application/json": {}
            },
            "description": "Referral not found"
          }
        },
        "summary": "Get the Action Plan summary information associated with a Referral",
        "tags": [
          "action-plan-controller"
        ]
      }
    },
...
    "/referral/{referralReference}/action-plan/action": {
      "post": {
        "operationId": "submitAction",
        "parameters": [
          {
            "in": "path",
            "name": "referralReference",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionPlanActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlanActionResponse"
                }
              }
            },
            "description": "Action submitted successfully"
          },
          "400": {
            "content": {
              "application/json": {}
            },
            "description": "Validation failure"
          },
          "404": {
            "content": {
              "application/json": {}
            },
            "description": "Referral not found"
          }
        },
        "summary": "Submit a need, outcome, and activities for an action plan",
        "tags": [
          "action-plan-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Development",
      "url": "https://community-support-api-dev.hmpps.service.justice.gov.uk"
    },
    {
      "description": "Pre-Production",
      "url": "https://community-support-api-preprod.hmpps.service.justice.gov.uk"
    },
    {
      "description": "Production",
      "url": "https://community-support-api.hmpps.service.justice.gov.uk"
    },
    {
      "description": "Local",
      "url": "http://localhost:8080"
    }
  ],
...
}