# src/main/kotlin/uk/gov/justice/digital/hmpps/communitysupportapi/controller/ActionPlanController.kt
package uk.gov.justice.digital.hmpps.communitysupportapi.controller

import io.swagger.v3.oas.annotations.Operation
import io.swagger.v3.oas.annotations.media.Content
import io.swagger.v3.oas.annotations.media.Schema
import io.swagger.v3.oas.annotations.responses.ApiResponse
import io.swagger.v3.oas.annotations.responses.ApiResponses
import jakarta.validation.Valid
import org.slf4j.LoggerFactory
import org.springframework.http.ResponseEntity
import org.springframework.security.access.prepost.PreAuthorize
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PatchMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RestController
import uk.gov.justice.digital.hmpps.communitysupportapi.authorization.UserMapper
import uk.gov.justice.digital.hmpps.communitysupportapi.dto.ActionPlanActionRequest
import uk.gov.justice.digital.hmpps.communitysupportapi.dto.ActionPlanActionResponse
import uk.gov.justice.digital.hmpps.communitysupportapi.dto.ActionPlanNeedsOrderRequest
import uk.gov.justice.digital.hmpps.communitysupportapi.dto.ActionPlanSelectANeedResponse
import uk.gov.justice.digital.hmpps.communitysupportapi.dto.ActionPlanSessionDeliveryDetailsRequest
import uk.gov.justice.digital.hmpps.communitysupportapi.dto.ActionPlanSessionDeliveryDetailsResponse
import uk.gov.justice.digital.hmpps.communitysupportapi.dto.ActionPlanSummaryDto
import uk.gov.justice.digital.hmpps.communitysupportapi.service.ActionPlanService
import uk.gov.justice.hmpps.kotlin.auth.HmppsAuthenticationHolder

@RestController
@PreAuthorize("hasAnyRole('ROLE_IPB_FRONTEND_RW')")
class ActionPlanController(
  private val actionPlanService: ActionPlanService,
  private val userMapper: UserMapper,
  private val authenticationHolder: HmppsAuthenticationHolder,
) {
  companion object {
    private val log = LoggerFactory.getLogger(this::class.java)
  }

  @Operation(summary = "Get the Action Plan summary information associated with a Referral")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Action Plan, and Referral, found - data returned",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSummaryDto::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @GetMapping("/bff/referral/{referralReference}/action-plan")
  fun getActionPlanSummary(@PathVariable referralReference: String): ResponseEntity<ActionPlanSummaryDto> {
    log.info("Fetching action plan summary for referral={}", referralReference)
    return ResponseEntity.ok(actionPlanService.getActionPlanSummaryForReferral(referralReference))
  }

  @Operation(summary = "Update the Action Plan needs order")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Action Plan needs order updated",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSummaryDto::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
      ApiResponse(
        responseCode = "500",
        description = "Unrecoverable error occurred",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @PatchMapping("/referral/{referralReference}/action-plan/update-needs-order")
  fun patchActionPlanNeedsOrder(
    @PathVariable referralReference: String,
    @Valid @RequestBody request: ActionPlanNeedsOrderRequest,
  ): ResponseEntity<ActionPlanSummaryDto> {
    log.info("Updating action plan needs order")
    return ResponseEntity.ok(actionPlanService.updateNeedsOrder(referralReference, request))
  }

  @Operation(summary = "Get the needs and outcomes for select a need")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Needs and outcomes returned",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSelectANeedResponse::class),
          ),
        ],
      ),
    ],
  )
  @GetMapping("/bff/referral/action-plan/select-a-need")
  fun getNeedsAndOutcomes(): ResponseEntity<ActionPlanSelectANeedResponse> {
    log.info("Fetching select-a-need data")
    return ResponseEntity.ok(actionPlanService.getNeedsAndOutcomesForActionPlan())
  }

  @Operation(summary = "Get the session delivery details with questions and saved answers for an action plan")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Session delivery details with questions and saved answered returned",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSessionDeliveryDetailsResponse::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @GetMapping("/bff/referral/{referralReference}/action-plan/session-delivery-details/session-delivery")
  fun getSessionDeliveryDetails(@PathVariable referralReference: String): ResponseEntity<ActionPlanSessionDeliveryDetailsResponse> {
    log.info("Fetching session delivery details for referral={}", referralReference)
    return ResponseEntity.ok(actionPlanService.getSessionDeliveryDetailsForReferral(referralReference))
  }

  @Operation(summary = "Save and update answers for session delivery details")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Session delivery details answers saved",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSessionDeliveryDetailsResponse::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "400",
        description = "Validation failure",
        content = [Content(mediaType = "application/json")],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @PatchMapping("/referral/{referralReference}/action-plan/session-delivery-details")
  fun patchSessionDeliveryDetails(
    @PathVariable referralReference: String,
    @Valid @RequestBody request: ActionPlanSessionDeliveryDetailsRequest,
  ): ResponseEntity<ActionPlanSessionDeliveryDetailsResponse> {
    val user = userMapper.fromToken(authenticationHolder)
    val changedBy = user.hmppsAuthUsername
    log.info("Saving session delivery details for referral={}", referralReference)
    return ResponseEntity.ok(actionPlanService.updateSessionDeliveryDetailsForActionPlan(referralReference, request, changedBy))
  }

  @Operation(summary = "Get the risk and adjustments questions and saved answers of the service delivery details")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Risk and adjustments questions and saved answers returned",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSessionDeliveryDetailsResponse::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @GetMapping("/bff/referral/{referralReference}/action-plan/service-delivery-details/risks-and-adjustments")
  fun getRiskAndAdjustments(@PathVariable referralReference: String): ResponseEntity<ActionPlanSessionDeliveryDetailsResponse> {
    log.info("Fetching service delivery risks and adjustments for referral={}", referralReference)
    return ResponseEntity.ok(actionPlanService.getRiskAndAdjustmentsForReferral(referralReference))
  }

  @Operation(summary = "Get the confirm service end date questions and saved answers of the service delivery details")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Confirm service end date questions and saved answers returned",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSessionDeliveryDetailsResponse::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @GetMapping("/bff/referral/{referralReference}/action-plan/service-delivery-details/confirm-service-end-date")
  fun getConfirmServiceEndDate(@PathVariable referralReference: String): ResponseEntity<ActionPlanSessionDeliveryDetailsResponse> {
    log.info("Fetching confirm service end date service delivery details for referral={}", referralReference)
    return ResponseEntity.ok(actionPlanService.getConfirmServiceEndDateForReferral(referralReference))
  }

  @Operation(summary = "Get the update service end date questions and saved answers of the service delivery details")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Update service end date questions and saved answers returned",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSessionDeliveryDetailsResponse::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @GetMapping("/bff/referral/{referralReference}/action-plan/service-delivery-details/update-service-end-date")
  fun getUpdateServiceEndDate(@PathVariable referralReference: String): ResponseEntity<ActionPlanSessionDeliveryDetailsResponse> {
    log.info("Fetching update service end date service delivery details for referral={}", referralReference)
    return ResponseEntity.ok(actionPlanService.getUpdateServiceEndDateForReferral(referralReference))
  }

  @Operation(summary = "Get the person involvement questions and saved answers of the service delivery details")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Person involvement questions and saved answers returned",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanSessionDeliveryDetailsResponse::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @GetMapping("/bff/referral/{referralReference}/action-plan/service-delivery-details/person-involvement")
  fun getPersonInvolvement(@PathVariable referralReference: String): ResponseEntity<ActionPlanSessionDeliveryDetailsResponse> {
    log.info("Fetching person involvement service delivery details for referral={}", referralReference)
    return ResponseEntity.ok(actionPlanService.getPersonInvolvementForReferral(referralReference))
  }

  @Operation(summary = "Submit a need, outcome, and activities for an action plan")
  @ApiResponses(
    value = [
      ApiResponse(
        responseCode = "200",
        description = "Action submitted successfully",
        content = [
          Content(
            mediaType = "application/json",
            schema = Schema(implementation = ActionPlanActionResponse::class),
          ),
        ],
      ),
      ApiResponse(
        responseCode = "400",
        description = "Validation failure",
        content = [Content(mediaType = "application/json")],
      ),
      ApiResponse(
        responseCode = "404",
        description = "Referral not found",
        content = [Content(mediaType = "application/json")],
      ),
    ],
  )
  @PostMapping("/referral/{referralReference}/action-plan/action")
  fun submitAction(
    @PathVariable referralReference: String,
    @Valid @RequestBody request: ActionPlanActionRequest,
  ): ResponseEntity<ActionPlanActionResponse> {
    val user = userMapper.fromToken(authenticationHolder)
    val changedBy = user.hmppsAuthUsername
    log.info("Submitting action for referral={} with need={}", referralReference, request.needId)
    return ResponseEntity.ok(actionPlanService.submitActionForReferral(referralReference, request, changedBy))
  }
}